[FIX] snailmail_account: use HttpCase to contact pingen
Pingen is a service that print (like really, using a printer) pdfs
document in order to mail them (using the real post and postmen). In
order to ensure we are compatible with their API, we send to their
sandbox environment a bunch of standard documents (like an invoice).

Those PDFs documents are generated our side using wkhtmltopdf but as the
test were not started using a HttpCase, the assets were not correctly
available.

This also reverts commit 55a2da45dd40460d6418f180fe067c75381a6749.

closes odoo/odoo#40323

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>