[FIX] account: duplicate tax when removing analytic
The method _get_tax_grouping_key_from_tax_line could give twice the same
key if we removed the analytic tags or tax ids: when 2 lines were
slightly different but are merged, a key conflict could happen.

If this is the case, drop the subsequent lines with the same key.

closes odoo/odoo#40583

X-original-commit: 7383452fbfc6718e32a46f5d727119171303211e
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>