[FIX] web_editor: remove destroyed editors from memory
Commit https://github.com/odoo/odoo/commit/0a25c4c920a20930142bad7098e0059bcd11914d
introduced a method to destroy all current editors to refresh them when
necessary... but forgot to empty the array that contains them.

In a 12.0 without custo, this is only slowing down the editor but in
master (or with custo) this makes the editor crash.

closes odoo/odoo#40372

X-original-commit: 2b5661937d9772b98c73f9f64aa419fec7676052
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>