[FIX] website: fix visitor _compute_time_statistics
This commit fixes the visitor's time_statistics computation. When creating a
new visitor, last connection datetime is not set, so it's not retrieved in the
search_read and crash at search_read_result[visitor.id].

This fix also speeds up and simplifies the time_statistics computation.
As time_connection_datetime is always set (for already created visitor) and in
the depends, no need to read values before looping, the data is already fetched
in memory. We can then use directly the value for each visitor in self.

Task ID: 2120464
PR #40199

closes odoo/odoo#43780

X-original-commit: d546c2e941ab2f9cef9bd3297f72702dd02cb36e
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>