[FIX] stock: display assign SN  immediate transfer
Before this commit, when user creates a receipt for a product tracked by
serial number, using immediate transfer, the fields to generate and
assign SN was hidden.
The reason was the move is confirmed in this case, and those fields was
displayed only for partially available and assigned moves.

task-2123761

X-original-commit: 4f9b170e857509cf34e072051086a7e9590c330d